backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
pydantic>=2.9.0
python-multipart>=0.0.12
scikit-learn>=1.6.0
nltk>=3.9.0
openai>=1.50.0
python-dotenv>=1.0.1
httpx>=0.27.0
groq>=0.11.0
